make cpu usage work on win
